pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "pre-commit>=4.2.0"
]

[project]
dependencies = [
  "bs4>=0.0.2",
  "graphcast[plot]>=0.14.2",
  "python-debianbts>=4.1.1",
  "suthing>=0.4.1",
  "tqdm>=4.67.1"
]
description = "Add your description here"
name = "deb-kg"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]
